Updated readme.  added model functions to get specific user data.
